Refactor calvaDebug into its own folder and util functions into separate file
