+++ src/extension-test/unit/calvaDebug/calvaDebug-test.ts
@@ -1,12 +1,12 @@
 import { expect } from 'chai';
-import { LispTokenCursor } from '../../../cursor-doc/token-cursor';
-import { getBreakpointTokenCursor } from '../../../calvaDebug';
-import * as vscode from 'vscode';
+import { moveTokenCursorToBreakpoint } from '../../../debugger/util';
+import * as mock from '../common/mock';
+import * as fs from "fs";
 
 describe('Calva Debug', async () => {
 
-    //const document = await vscode.workspace.openTextDocument('../../../../test-data/calva-debug.clj');
     let debugResponse: any;
+    let doc: mock.MockDocument;
 
     beforeEach(() => {
         debugResponse = {
@@ -15,11 +15,16 @@ describe('Calva Debug', async () => {
         };
     });
 
-    describe('getBreakpointTokenCursor', () => {
+    describe('moveTokenCursorToBreakpoint', () => {
 
-        it('should return token cursor at correct offset', async () => {
-            //const document = await vscode.workspace.openTextDocument('test-data.clj');
-            //const tokenCursor = getBreakpointTokenCursor(document, debugResponse);
-        });
+        // it('should return token cursor at correct offset', async () => {
+        //     const path = __dirname + '/test-code.clj';
+        //     const docText = fs.readFileSync(path, "utf8");
+        //     doc.insertString(docText);
+        //     const tokenCursor = doc.getTokenCursor(0);
+        //     debugResponse.coor = [3, 2];
+        //     moveTokenCursorToBreakpoint(tokenCursor, debugResponse);
+        //     expect(tokenCursor.getPrevToken().raw).equal('x');
+        // });
     });
 });
