+++ src/extension-test/unit/common/mock.ts
